+++ lib/reactiv.dart
@@ -5,10 +5,4 @@ export 'dependency_management/dependency.dart';
 export 'state_management/reactive_types.dart';
 export 'state_management/widgets/observer.dart';
 export 'utils/logger.dart';
-export 'views/reactive_view.dart';
-
-// /// A Calculator.
-// class Calculator {
-//   /// Returns [value] plus 1.
-//   int addOne(int value) => value + 1;
-// }
+export 'views/reactive_widget.dart';