+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.3
+1. Stream change to broadcast stream,
+2. params change
+   1. listen => listenable
+   2. update => listener
+3. ReactiveWidget major update :
+   1. auto dispose the controller functionality
+   2. bindDependency method to Dependency.put() the dependency
+   3. life-cycle methods void initState(), void dispose(), void onBuildCompleted(BuildContext context), methods added in ReactiveWidget
+
 ## 0.1.2
 1. [dot]reactiv extension added
 2. ReactiveList all the methods added for add new element and remove element from the list